refactor(datasource): split wecom archive client by build target
